OK this one works I swear

+++ deploy_site.sh
@@ -2,6 +2,8 @@
 
 printf "\033[0;32m[[=== Deploying Site ===]]\033[0;34m\n"
 
+ARG="$1"
+
 # Make it slightly harder for scrapers to get internal information.
 DEPLOY='c2Z0cCAtb1VzZXI9c3RyaWZmYSBzZnRwLnJlZWQuZWR1Cg=='
 DEPLOY_CMD="`base64 -d <<< $DEPLOY`"
@@ -9,20 +11,24 @@ printf "\033[0;35m${DEPLOY_CMD}\033[0;34m\n"
 
 MANIFEST="manifest.txt"
 
-(
-rm -f "$MANIFEST";
-echo "get html/$MANIFEST";
-sleep 0.2
-while [ ! -f "$MANIFEST" ]; do
-	sleep 0.5;
-done;
-md5sum -c --quiet "$MANIFEST" \
-	| cut --complement -d : -f 2- \
-	| awk '{print "put -f " $1;}';
-find html -type f -print0 \
-	| xargs -0 md5sum -b > "$MANIFEST";
-echo "put $MANIFEST html";
-) | tee sftp.log | sh -c "$DEPLOY_CMD"
+if [[ "$ARG" = "force" ]]; then
+	echo 'put -rf html';
+else
+	rm -f "$MANIFEST";
+	echo "get html/$MANIFEST html/$MANIFEST";
+	find html -type f -print0 \
+		| xargs -0 md5sum -b > "$MANIFEST";
+	while [ ! -f "html/$MANIFEST" ]; do
+		sleep 0.5;
+	done;
+	cat "$MANIFEST" "html/$MANIFEST" \
+		| sort \
+		| uniq -u\
+		| cut --complement -d '*' -f -1 \
+		| sort -u \
+		| awk '{print "put -f " $1;}';
+	echo "put $MANIFEST html";
+fi | tee sftp.log | sh -c "$DEPLOY_CMD"
 
 printf "\033[0;32m[[=== Site Deployed. ===]]\033[0m\n"
